tests/wpt/web-platform-tests/css/css-inline/parsing/line-height-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Inline Layout: parsing line-height with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-inline-3/#line-height-property">
<meta name="assert" content="line-height supports only the grammar 'normal | <number> | <length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("line-height", "auto");

test_invalid_value("line-height", "-2");
test_invalid_value("line-height", "-10px");
test_invalid_value("line-height", "-200%");

test_invalid_value("line-height", "2 10px");
test_invalid_value("line-height", "200% 3");
test_invalid_value("line-height", "auto 10px");
test_invalid_value("line-height", "3 auto");
</script>
</body>
</html>
